requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
Pillow>=10.2.0,<11.0.0
pillow-heif>=0.14.0,<1.0.0
pillow-avif-plugin>=1.4.0,<2.0.0
pyoxipng>=9.0.0,<10.0.0
scour>=0.38.0,<1.0.0
python-multipart>=0.0.6,<1.0.0
httpx>=0.27.0,<1.0.0
google-cloud-storage>=2.14.0,<3.0.0
defusedxml>=0.7.0,<1.0.0
redis[hiredis]>=5.0.0,<6.0.0
pydantic-settings>=2.1.0,<3.0.0
scipy>=1.13,<2.0
